+++ packages/api/test/integration/queries.test.ts
@@ -2,6 +2,7 @@ import { execute, parse } from 'graphql'
 import { beforeEach, expect, test, vi } from 'vitest'
 
 import { GraphqlVtexContextFactory, GraphqlVtexSchema } from '../../src'
+import type { Options } from '../../src/typings/globals'
 import {
   AllCollectionsQueryFirst5,
   catalogBrandListFetch,